Split and prune VLM evaluation tests
